`buddypress` post type dynamic creation for `has_directory` components
12.0 introduces a `buddypress` post type which is used to store the components directory pages into the database & replace the `page` post type that was used so far. It also replaced the Page association WP Admin screen with another WP Admin screen to customize all BP URLs.

In case a custom component is using a directory page and is not taking care of generating the corresponding `buddypress` post type entry, instead of asking for a page association in an admin notice, this entry is dynamically generated. If users are not happy with the title or slug of this entry, they can use the BP URLs settings tab to customize these afterwards.

Fixes #8918
See #4954
Closes https://github.com/buddypress/buddypress/pull/116



git-svn-id: https://buddypress.svn.wordpress.org/trunk@13502 cdf35c40-ae34-48e0-9cc9-0c9da1808c22