Add a function to output  the `bp-activity-comment-id` data attribute
In BP Nouveau, instead of using the same function to output the `bp-activity-comment-id` or `bp-activity-id` data attribute according to the activity type, introduce the `bp_nouveau_activity_comment_data_attribute_id()` function and use it inside the `activity/comment.php` template.
This prevents a notice warning notice and avoid to query for the parent activity to get the right activity type.

Props vapvarun

Fixes #9014
Closes https://github.com/buddypress/buddypress/pull/182



git-svn-id: https://buddypress.svn.wordpress.org/trunk@13631 cdf35c40-ae34-48e0-9cc9-0c9da1808c22