BP Blocks: make sure default block titles can be translated
Some of the BP Blocks are using a `title` attribute into their `block.json`. Unfortunately, block attributes do not seem to be include in the translatable strings of the WP Block Metadata API.

To fix this, all `block.json` `title` attributes now default to an empty string & the default title value is handled within the PHP callback to render the block.

For the specific case of the Latest Activities block, this commit also fixes some errors that were preventing it to be able to select the list of Activity types to include into the rendered activities.

See #9074 (trunk)
Closes https://github.com/buddypress/buddypress/pull/223



git-svn-id: https://buddypress.svn.wordpress.org/trunk@13716 cdf35c40-ae34-48e0-9cc9-0c9da1808c22