WPCS: fix a regression that stopped the BuddyPress components from being saved properly.
The regression was introduced as part of [13881] where `"bp_components[ $name ]"` was incorectly changed (added a space around it). The values sent to the database had a space before and after the component key: `" groups "` instead of `"groups"`, and it was saved in the `bp-active-components` field as such.

The correct value should continue to be `"bp_components[$name]"`, since this is not an array key (even though it looks like one).

See #7228

git-svn-id: https://buddypress.svn.wordpress.org/trunk@13906 cdf35c40-ae34-48e0-9cc9-0c9da1808c22